refactor(composites): flatten + dedupe the nested mock config loader; drop dead imports
Four composite generators (baseline, departitioned, reconciled,
millard_pdmp_baseline) each defined their own nested _CachedLoader with an
identical 4-level inner-class mock (_SimData -> _InternalState ->
_UniqueMolecule). Replaced all four with a single flattened, module-level
CachedConfigLoader in _helpers.py (no class-in-class nesting). Also removed two
dead imports in chromosome_structure.py (GlobalClock, UniqueUpdate from the old
ecoli.processes namespace — never referenced).

Behavior-preserving: all four architectures build; baseline 20-tick cell_mass is
byte-identical (1266.952616). Net -60 lines.
