+++ v2ecoli/composites/reconciled.py
@@ -35,6 +35,7 @@ from v2ecoli.composites._helpers import (
     _normalize_boundary_units,
     _make_instance,
     _get_special_step,
+    CachedConfigLoader,
     _expand_flushes,
     FLUSH,
     PARTITIONED_PROCESSES,
@@ -430,35 +431,8 @@ def reconciled(core: Any = None, *, seed: int = 0, cache_dir: str = "out/cache")
     })
 
     # Mock loader
-    class _CachedLoader:
-        def __init__(self, configs, unique_names, dry_mass_inc_dict):
-            self._configs = configs
-            self.unique_names = unique_names
-
-            class _SimData:
-                class _InternalState:
-                    class _UniqueMolecule:
-                        def __init__(self, names):
-                            self.unique_molecule_definitions = {
-                                n: {} for n in names}
-                    unique_molecule = None
-                    def __init__(self, names):
-                        self.unique_molecule = self._UniqueMolecule(names)
-                internal_state = None
-                expectedDryMassIncreaseDict = {}
-
-            self.sim_data = _SimData()
-            self.sim_data.internal_state = _SimData._InternalState(
-                unique_names)
-            self.sim_data.expectedDryMassIncreaseDict = (
-                dry_mass_inc_dict or {})
-
-        def get_config_by_name(self, name):
-            if name in self._configs:
-                return self._configs[name]
-            raise KeyError(f'Unknown: {name}')
-
-    loader = _CachedLoader(configs, unique_names, dry_mass_inc_dict)
+    # Mock loader: cache configs + minimal sim_data (see _helpers.CachedConfigLoader).
+    loader = CachedConfigLoader(configs, unique_names, dry_mass_inc_dict)
 
     # Build execution layers for the default feature set
     execution_layers = build_execution_layers(DEFAULT_FEATURES)
